// src/app/models/admin/capphephoatdongkhoangsan/cpthamdotoadokhuvuc.model.ts
export class OutputCpThamDoToaDoKhuVucModel {
    idthamdotoadokhuvuc: string;
    sohieu: string;
    toadox: string;
    toadoy: string;
    idcapphepthamdo: string;
    idthamdokhuvuc: string;
    loaikhuvuc: number;
    loaicapphep: number;
    siteid: string;
    thutu: number
}

export class InputCpThamDoToaDoKhuVucModel {
    idthamdotoadokhuvuc: string;
    sohieu: string;
    toadox: string;
    toadoy: string;
    idcapphepthamdo: string;
    idthamdokhuvuc: string;
    loaicapphep: number;
    loaikhuvuc: number;
    thutu: number
}