// src/app/models/admin/capphephoatdongkhoangsan/cpthamdokhuvuc.model.ts
export class OutputCpThamDoKhuVucModel {
    idthamdokhuvuc: string;
    tenkhuvuc: string;
    dientich: number;
    donvidientich: string;
    idcapphepthamdo: string;
    loaikhuvuc: number;
    geom: string;
    geovn: string;
    geowgs: string;
    hequychieu: string;
    siteid: string;
}

export class InputCpThamDoKhuVucModel {
    idthamdokhuvuc: string;
    tenkhuvuc: string;
    dientich: number;
    donvidientich: string;
    idcapphepthamdo: string;
    loaikhuvuc: number;
    hequychieu: string;
    loaicapphep: string;
    listtoado = [];
}